refactor: create use logs hook, simplify step-run-logs.tsx
